# doc/api.yml
openapi: 3.0.3
info:
  title: Birdia API
  description: |
    Detect and qualify geo-referenced objects.
    
    # Versioning
    This API uses semantic versioning (MAJOR.MINOR.PATCH). Available versions are : 
    - [v1.3.2 (2026-02-25)](https://petstore.swagger.io/?url=https://raw.githubusercontent.com/b-partners/geo-jobs/prod/doc/version/v1-3-2-api.yml)
    - [v1.3.1 (2026-02-20)](https://petstore.swagger.io/?url=https://raw.githubusercontent.com/b-partners/geo-jobs/prod/doc/version/v1-3-1-api.yml) *(current)*
    - [v1.3.0 (2026-02-16)](https://petstore.swagger.io/?url=https://raw.githubusercontent.com/b-partners/geo-jobs/prod/doc/version/v1-3-api.yml) 
    - [v1.2.0 (2025-12-10)](https://petstore.swagger.io/?url=https://raw.githubusercontent.com/b-partners/geo-jobs/prod/doc/version/v1-2-api.yml) 
    - [v1.1.0 (2025-05-13)](https://petstore.swagger.io/?url=https://raw.githubusercontent.com/b-partners/geo-jobs/prod/doc/version/v1-1-api.yml)
    - [v1.0.0 (2024-10-25)](https://petstore.swagger.io/?url=https://raw.githubusercontent.com/b-partners/geo-jobs/prod/doc/version/v1-0-api.yml) *(deprecated)*
    
    # Security
    
    Get your **API key** by emailing **contact@birdia.fr**.
    All of your requests must embed your API key in the header named: **x-api-key**.
    
    # Glossary
    - *Detection* : process that involves analyzing images of a geographic area using an AI model. 
...
    
...
    
...
    
...
    
...
    
...
    
...
    
    # Usage
    
...
    
...
    
    # Detection examples
...

...
servers:
...
paths:
...
  '/image':
    get:
      tags:
        - Image
      summary: Retrieve an aerial image from a human-readable address or geographic coordinates.
      description: |
        Retrieve an aerial image by providing either a human-readable address
        or a pair of geographic coordinates (longitude and latitude). 
        
        When `ROOF_DAMAGE_DETECTION` is provided, only images with 5cm precision is returned.
      operationId: getImage
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: query
          name: usage
          schema:
            $ref: '#/components/schemas/ImageUsage'
          required: false
        - in: query
          name: address
          schema:
            type: string
          required: false
        - in: query
          name: longitude
          schema:
            type: number
          required: false
        - in: query
          name: latitude
          schema:
            type: number
          required: false
        - in: query
          name: zoom
          schema:
            $ref: '#/components/schemas/ImageZoomLevel'
          required: false
          description: Default value is BUILDING.
        - in: query
          name: isExtended
          schema:
            type: boolean
            default: true
          required: false
          deprecated: true
        - in: query
          name: shiftNb
          schema:
            type: integer
            default: 0
          required: false
      responses:
        200:
          description: Image details of provided address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDetails'
...
  '/geocode':
    get:
      tags:
        - Geocode
      operationId: geocodeAddress
      summary: Resolve an address into a GeoJSON Feature
      description: |
        Converts a textual address into a geographic location and returns the result as a GeoJSON Feature.
        The service performs forward geocoding by querying external geocoding providers and transforms the response into a standardized GeoJSON Feature object.
      parameters:
        - in: query
          name: address
          required: true
          description: Address to geocode
          schema:
            type: string
      responses:
        200:
          description: GeoJSON Feature corresponding to address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feature'
  '/geoCodingJobs/{id}':
    get:
      tags:
        - Geocode
      operationId: retrieveGeoCodingJobById
      summary: Retrieve a geo-coding job resolving list of addresses to GeoJSON features
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: GeoCodingJob with its status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoCodingJob'
  '/geoCodingJobs/{id}/excel':
    post:
      tags:
        - Geocode
      operationId: geocodeAddressesThroughExcelAddresses
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      summary: Resolve addresses contain inside provided excel file to GeoJSON Features
      description: |
        Converts list textual addresses inside provided excel file into a geographic location and returns the result as list of GeoJSON features.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Created GeoCodingJob
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoCodingJob'
components:
  securitySchemes:
    ReadmeAuth:
      type: apiKey
      name: readme-signature
      in: header
...
  schemas:
...
    ImageZoomLevel:
      type: string
      enum:
        - BUILDING
        - NEIGHBORHOOD
        - DISTRICT
...
    GeoCodingJob:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/Status'
        geoJsonUrl:
          type: string
...
    TileCoordinates:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer
        z:
          type: integer
...
    Status:
      type: object
      properties:
        progression:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - FINISHED
        health:
          type: string
          enum:
            - SUCCEEDED
            - FAILED
            - UNKNOWN
            - RETRYING
        creationDatetime:
          type: string
          format: date-time

    ###########################################################################################################
    ###  GeoJSON objects. Modified from: https://gist.github.com/bubbobne/fe5f2db65acf039be6a9fd92fc9c7233  ###
    ###########################################################################################################
...
    Feature:
      type: object
      properties:
        type:
          type: string
          enum: [ Feature ]
        geometry:
          oneOf:
            - $ref: '#/components/schemas/Point'
            - $ref: '#/components/schemas/Polygon'
            - $ref: '#/components/schemas/MultiPolygon'
        properties:
          type: object
          additionalProperties: true
    Geometry:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - Point
            - Polygon
            - MultiPolygon
    Point2D:
      type: array
      maxItems: 2
      minItems: 2
      items:
        type: number
    Point:
      type: object
      allOf:
        - $ref: "#/components/schemas/Geometry"
      properties:
        coordinates:
          $ref: '#/components/schemas/Point2D'
    Polygon:
      type: object
      allOf:
        - $ref: "#/components/schemas/Geometry"
      properties:
        coordinates:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/Point2D'
    MultiPolygon:
      type: object
      allOf:
        - $ref: "#/components/schemas/Geometry"
      properties:
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                $ref: '#/components/schemas/Point2D'
...
    ImageDetails:
      allOf:
        - $ref: '#/components/schemas/ImageOfAddress'
      properties:
        minTileCoordinates:
          $ref: '#/components/schemas/TileCoordinates'
        currentGeoPosition:
          $ref: '#/components/schemas/GeoPosition'
        zoomLevel:
          $ref: '#/components/schemas/ImageZoomLevel'
    GeoPosition:
      type: object
      properties:
        longitude:
          type: number
        latitude:
          type: number
    ImageOfAddress:
      type: object
      deprecated: true
      properties:
        address:
          type: string
        imageBase64:
          type: string
...
    ImageUsage:
      type: string
      enum:
        - ROOF_DAMAGE_DETECTION
        - UNKNOWN